Update README.md
I have basically created a tree type structure using c++. Here I have taken the root node with pointers to owner name, address, mobile, phone, value, and a nodenumber which tells the node at which we are. and the nodenumber furthe point to a newnode and hence user can add values to it. I just want to say I have tried to code my logic in a best way. There can be a few compile errors as there was a little short of time but my logic is clear.